changed sending likes to forms

+++ src/ViazushkiBundle/Controller/LikeController.php
@@ -2,47 +2,52 @@
 
 namespace ViazushkiBundle\Controller;
 
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Component\HttpFoundation\Request;
 use ViazushkiBundle\Entity\Like;
+use ViazushkiBundle\Entity\Toy;
+use ViazushkiBundle\Form\Type\LikeType;
 
 class LikeController extends Controller
 {
-    public function addAction($toyId)
+    /**
+     * @ParamConverter("toy", class="ViazushkiBundle:Toy")
+     */
+    public function updateAction(Request $request, Toy $toy)
     {
+        if (!$toy) {
+            throw $this->createNotFoundException('Страница не найдена');
+        }
+
         $em = $this->getDoctrine()->getManager();
 
-        $toy = $em->getRepository('ViazushkiBundle:Toy')->find($toyId);
-        if (!$toy) {
-            throw $this->createNotFoundException();
-        }
         $like = new Like();
+        $likeForm = $this->createForm(LikeType::class);
+        $currentUser = $this->getUser();
 
-        $like
-            ->setUser($this->getUser())
-            ->setToy($toy)
-        ;
+        $likeForm->handleRequest($request);
+        if ($likeForm->isSubmitted() && $likeForm->isValid()) {
+            $like
+                ->setToy($toy)
+                ->setUser($currentUser)
+            ;
 
-        $em->persist($like);
-        $em->flush();
+            if (count($this->get('validator')->validate($like)) == 0) {
+                $em->persist($like);
+                $em->flush();
+            } else {
+                $like = $em->getRepository('ViazushkiBundle:Like')->findOneBy([
+                   'toy' => $toy,
+                   'user' => $currentUser,
+                ]);
 
-        return $this->redirectToRoute('viazushki_homepage');
-    }
-
-    public function deleteAction($toyId)
-    {
-        $em = $this->getDoctrine()->getManager();
-
-        $toy = $em->getRepository('ViazushkiBundle:Toy')->find($toyId);
-        if (!$toy) {
-            throw $this->createNotFoundException();
+                $em->remove($like);
+                $em->flush();
+            }
+            return $this->redirectToRoute('viazushki_toy', ['slug' => $toy->getSlug()]);
         }
-        $likes = $em->getRepository('ViazushkiBundle:Like')->findBy(['toy' => $toy->getId(), 'user' => $this->getUser()]);
 
-        foreach ($likes as $like) {
-            $em->remove($like);
-        }
-        $em->flush();
-
-        return $this->redirectToRoute('viazushki_homepage');
+        return $this->createNotFoundException('Страница не найдена');
     }
 }
